txscript: Move RuntimeSigOpCounter to separate module, improve docs
