+++ crypto/txscript/src/lib.rs
@@ -12,6 +12,8 @@ pub mod standard;
 #[cfg(feature = "wasm32-sdk")]
 pub mod wasm;
 
+pub mod runtime_sig_op_counter;
+
 use crate::caches::Cache;
 use crate::data_stack::{DataStack, Stack};
 use crate::opcodes::{deserialize_next_opcode, OpCodeImplementation};
@@ -30,6 +32,7 @@ use script_class::ScriptClass;
 pub mod prelude {
     pub use super::standard::*;
 }
+use crate::runtime_sig_op_counter::{RuntimeSigOpCounter, SigOpConsumer};
 pub use standard::*;
 
 pub const MAX_SCRIPT_PUBLIC_KEY_VERSION: u16 = 0;
@@ -72,80 +75,6 @@ pub struct SigCacheKey {
 enum ScriptSource<'a, T: VerifiableTransaction> {
     TxInput { tx: &'a T, input: &'a TransactionInput, idx: usize, utxo_entry: &'a UtxoEntry, is_p2sh: bool },
     StandAloneScripts(Vec<&'a [u8]>),
-}
-
-/// RuntimeSigOpCounter represents the state tracking of signature operations during script execution.
-/// Unlike the static counting approach which counts all possible signature operations,
-/// this tracks only the actually executed signature operations, leading to more accurate
-/// mass calculations and potentially lower fees for conditional scripts.
-#[derive(Debug, Clone)]
-pub struct RuntimeSigOpCounter {
-    /// Maximum number of signature operations allowed for this input
-    sig_op_limit: u8,
-    /// Remaining signature operations that can be executed
-    sig_op_remaining: u8,
-}
-
-impl RuntimeSigOpCounter {
-    pub fn new(sig_op_limit: u8) -> Self {
-        Self { sig_op_limit, sig_op_remaining: sig_op_limit }
-    }
-    /// Attempts to consume the specified number of signature operations.
-    ///
-    /// This method handles:
-    /// - Checking if we have enough remaining operations
-    /// - Updating the remaining count if successful
-    ///
-    /// # Arguments
-    /// * `num_sigs` - The number of signature operations to consume
-    ///
-    /// # Returns
-    /// * `Ok(())` if the operations were successfully consumed
-    /// * `Err(TxScriptError::ExceededSigOpLimit)` if not enough operations remain
-    ///
-    /// # Example
-    /// ```
-    /// let mut counter = kaspa_txscript::RuntimeSigOpCounter::new(1);
-    ///
-    /// // Consume 3 operations
-    /// counter.consume_sig_op().unwrap(); // Ok(())
-    /// assert_eq!(counter.sig_op_remaining(), 0);
-    ///
-    /// // Try to consume too many
-    /// counter.consume_sig_op().unwrap_err(); // Err(ExceededSigOpLimit)
-    /// ```
-    pub fn consume_sig_op(&mut self) -> Result<(), TxScriptError> {
-        self.sig_op_remaining = self.sig_op_remaining.checked_sub(1).ok_or(TxScriptError::ExceededSigOpLimit(self.sig_op_limit))?;
-
-        Ok(())
-    }
-
-    pub fn sig_op_remaining(&self) -> u8 {
-        self.sig_op_remaining
-    }
-
-    pub fn sig_op_limit(&self) -> u8 {
-        self.sig_op_limit
-    }
-}
-
-pub trait SigOpConsumer {
-    fn consume_sig_op(&mut self) -> Result<(), TxScriptError>;
-}
-
-impl SigOpConsumer for RuntimeSigOpCounter {
-    fn consume_sig_op(&mut self) -> Result<(), TxScriptError> {
-        RuntimeSigOpCounter::consume_sig_op(self)
-    }
-}
-impl SigOpConsumer for Option<RuntimeSigOpCounter> {
-    fn consume_sig_op(&mut self) -> Result<(), TxScriptError> {
-        if let Some(consumer) = self {
-            consumer.consume_sig_op()
-        } else {
-            Ok(())
-        }
-    }
 }
 
 pub struct TxScriptEngine<'a, T: VerifiableTransaction, Reused: SigHashReusedValues> {
@@ -310,14 +239,11 @@ impl<'a, T: VerifiableTransaction, Reused: SigHashReusedValues> TxScriptEngine<'
         }
     }
 
-    /// Returns the number of signature operations that were actually required during script execution.
-    /// Must only be called after script execution completes successfully.
+    /// Returns the number of signature operations used in script execution if runtime sig op counting is enabled.
     ///
-    /// Returns the difference between the input's sig_op_limit and remaining sig ops.
+    /// Returns None if runtime signature operation counting is disabled.
     pub fn used_sig_ops(&self) -> Option<u8> {
-        self.runtime_sig_op_counter
-            .as_ref()
-            .map(|RuntimeSigOpCounter { sig_op_limit, sig_op_remaining }| sig_op_limit - sig_op_remaining)
+        self.runtime_sig_op_counter.as_ref().map(|counter| counter.used_sig_ops())
     }
 
     /// Creates a new Script Engine for validating transaction input.
